Implemented detect anomalies of speeding
Removed WINDOW w1 as it cannot be used for calculating speed of the last user
Removed CalculateSpeed, VoltTrigger on 'w1'

Added step to Remove old anomalies before inserting new one

Added step to calculate speed in ProcessBikestatus
Added code copied from Web to calculate distance between two GEO points (necessary for calculating speed in mph)

Added the 'time' field to the 'riderPositions' TABLE to use as previous location for calculating speed
Changed the 'speed' field data type to FLOAT in the 's2' STREAM

+++ src/benchmarks/edu/brown/benchmark/bikerstream/BikerStreamConstants.java
@@ -180,7 +180,7 @@ public abstract class BikerStreamConstants {
     // ANOMALIES
     //
 
-    public static final long STOLEN_SPEED = 60;
-    public static final long STOLEN_STATUS = 1;
+    public static final double STOLEN_SPEED = 60.0;
+    public static final long   STOLEN_STATUS = 1;
 }
